+++ src/app/locales/en/components/space-automation/input-stores-form.js
@@ -1,22 +1,7 @@
 export default {
   cannotGuessStoreForSelection: 'There is more than one input store suitable for the selected item(s).<br/>You can apply your selection to desired stores with the <strong>Use selected items</strong> button.',
-  fields: {
-    inputStores: {
-      inputStore: {
-        rawValue: {
-          errors: {
-            badValue: 'Provided value is not suitable for this store',
-          },
-        },
-      },
-    },
-  },
   useSelectionButton: {
     buttonText: 'Use selected items ({{itemsCount}})',
   },
-  fileTag: {
-    unknownName: 'Unknown',
-    idTooltip: 'ID: {{id}}',
-  },
-  filesSelectorSubheader: 'for <strong>{{storeName}}</strong> store',
+  itemsSelectorSubheader: 'for <strong>{{storeName}}</strong> store',
 };