+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-10367 Improved editor of input stores content in workflow execution wizard
+* VFS-10365 Improved editor of automation store default value
 * VFS-9999 Added detailed space details popovers with space information in various views
 * VFS-10486 Removed custom zoom in tests rendering
 * VFS-10339 Added class-based implementation of a chart point